(PC-12635) feat(SetBirthday): rearrangement of components
