+++ src/features/auth/signup/SetBirthday/DatePicker/DatePickerSpinner.web.tsx
@@ -4,7 +4,7 @@ import Picker from 'react-mobile-picker'
 import styled from 'styled-components/native'
 import { v4 as uuidv4 } from 'uuid'
 
-import { DateInput } from 'features/auth/signup/SetBirthday/DateInput/DateInput'
+import { DateInput } from 'features/auth/signup/SetBirthday/atoms/DateInput/DateInput'
 import { useDatePickerErrorHandler } from 'features/auth/signup/SetBirthday/utils/useDatePickerErrorHandler'
 import { SignupData } from 'features/auth/signup/types'
 import {
