+++ src/features/cheatcodes/pages/AppComponents/AppComponents.tsx
@@ -7,8 +7,8 @@ import styled from 'styled-components/native'
 
 import { CallToActionIcon, CategoryIdEnum, PopOverIcon, VenueTypeCodeKey } from 'api/gen/api'
 import { SIGNUP_NUMBER_OF_STEPS } from 'features/auth/api'
-import { DateInput } from 'features/auth/signup/SetBirthday/DateInput/DateInput'
-import { DropDown } from 'features/auth/signup/SetBirthday/DatePicker/DropDown'
+import { DateInput } from 'features/auth/signup/SetBirthday/atoms/DateInput/DateInput'
+import { DropDown } from 'features/auth/signup/SetBirthday/atoms/DropDown/DropDown'
 import { EndedBookingTicket } from 'features/bookings/components/EndedBookingTicket'
 import { OnGoingTicket } from 'features/bookings/components/OnGoingTicket'
 import { ThreeShapesTicket } from 'features/bookings/components/ThreeShapesTicket'
